Moving mpris structures to their own file
